+++ packages/orq-rc/RELEASES.md
@@ -10969,3 +10969,13 @@ Based on:
 - [python v4.15.0-rc.50] packages/orq-rc
 ### Releases
 - [PyPI v4.15.0-rc.50] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.50 - packages/orq-rc
+
+## 2026-09-15 13:27:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.797.0 (2.937.18) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.15.0-rc.51] packages/orq-rc
+### Releases
+- [PyPI v4.15.0-rc.51] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.51 - packages/orq-rc